+++ src/App.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Header from './components/Header';
 import Home from './components/Home';
-import Company from './components/Company';
-import References from './components/References';
-import Partners from './components/Partners';
-import Contact from './components/Contact';
+import Company from './pages/Company';
+import References from './pages/References';
+import Partners from './pages/Partners';
+import Contact from './pages/Contact';
 import Footer from './components/Footer';
 
 const App = () => {
