--- src/components/Contact.css
@@ -1,34 +1,0 @@
-.contact-section {
-    display: flex;
-    justify-content: space-between;
-  }
-  
-  .left-section {
-    width: 60%;
-    padding: 20px;
-  }
-  
-  .right-section {
-    width: 35%;
-    padding: 20px;
-  }
-  
-  .map-container {
-    margin-bottom: 20px;
-    width: 100%;
-  }
-  
-  .contact-form .form-group {
-    margin-bottom: 10px;
-  }
-  
-  .btn-primary {
-    background-color: #007bff;
-    color: #fff;
-    padding: 10px 20px;
-    border: none;
-    cursor: pointer;
-  }
-  
-  /* Add more styles as needed */
-
