+++ src/pages/Contact.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Fade } from 'react-reveal';
-import './Contact.css'; // Import the CSS file for styling
+import '../styles/Contact.css'; // Import the CSS file for styling
 
 const Contact = () => {
   return (
@@ -15,7 +15,7 @@ const Contact = () => {
                 src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3322.662874476876!2d-6.855563385253903!3d33.97841322931239!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0xda6f6e77433a56f%3A0xc7fb3a9e6b1e4de7!2sR%C3%A9sidence%20Essabah%20Immeuble%20Boujdour%20CYM%2C%20Rabat!5e0!3m2!1sen!2sma!4v1628048803162!5m2!1sen!2sma"
                 width="100%"
                 height="450"
-                style={{ border: 0 }}
+                style={{ padding: 10 }}
                 allowFullScreen=""
                 loading="lazy"
               ></iframe>
@@ -26,6 +26,7 @@ const Contact = () => {
                   <label htmlFor="name">Nom</label>
                   <input type="text" className="form-control" id="name" name="name" required />
                 </div>
+                
                 <div className="form-group">
                   <label htmlFor="email">Email</label>
                   <input type="email" className="form-control" id="email" name="email" required />
@@ -48,6 +49,7 @@ const Contact = () => {
                     En envoyant mes informations, j'accepte votre Politique de confidentialité
                   </label>
                 </div>
+
                 <div>
                   <button type="submit" className="btn btn-primary">Envoyer</button>
                 </div>
